Introduce a IdentityManager trait
This acts as a way to unify behaviour between the initial legacy
contract and the new batching contract. It will allows us to integrate
the new one in parallel with the old and increase the maintainability of
the codebase while operating in parallel by reducing the need for
divergent implementations of core logic.

It does not integrate the new contract at this stage, and attempting to
use the `batching-contract` feature will result in a panic.