Implement interactions with the batching contract.
This commit removes any use of the legacy contract from the codebase.
This means that, while the code still exists due to some dependencies,
it is no longer used. The batching contract is no-longer gated behind a
feature flag.

It is now fully integrated with the batching contract on chain, as well
as the prover service used to generate proofs for verification on chain.
These interactions are tested, and while multiple components are mocked
the signup sequencer portion of the code is as it should be with the
real external dependencies.