Have the prover mock verify inputs
This commit ensures that the mock prover service now actually performs
verification on the provided proofs. This means that only a malicious
signup sequencer should be able to spuriously pass the integration
tests.

In addition, it ensures that the return value from the prover mock (the
proof) is now correlated with the input in a way that can be verified on
the mock chain.