Update the existing tests to use the new ABI
This commit updates the signup sequencer's integration tests to use the
new contract ABI corresponding to commit `04daa6d` in the
`world-id-contracts` repository.

It also fixes some outstanding clippy lints that had not yet been
cleaned up on `batching/main` and adds on-chain proof verification for
semaphore proofs to the tests.